+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed dependency `getrandom` ([rand#1674])
 - Removed optional dependency `serde` ([#28])
 - Add `SeedableRng::fork` methods ([#17])
+- Rename trait `block::BlockRngCore` to `block::Generator` and associated type `Results` to `Output`; remove assoc. `type Item` and remove type bounds ([#26])
 ### Other
 - Changed repository from [rust-random/rand] to [rust-random/core].
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [rand#1669]: https://github.com/rust-random/rand/pull/1669
 [rand#1674]: https://github.com/rust-random/rand/pull/1674
 [#17]: https://github.com/rust-random/rand-core/pull/17
+[#26]: https://github.com/rust-random/rand-core/pull/28
 [#28]: https://github.com/rust-random/rand-core/pull/28
 
 [rust-random/rand]: https://github.com/rust-random/rand